+++ spec/support/shared/an_active_record_matcher.rb
@@ -12,10 +12,14 @@ shared_examples_for 'an Active Record matcher' do
     positive,
     negative,
     context={},
-    macro_option={ option_name => positive }
+    macro_option=nil,
+    &block
   )
     context "with option #{option_name.inspect}" do
-      let(:options) { context.merge(macro_option) }
+      let(:option_name) { option_name }
+      let(:value) { positive }
+
+      let(:options) { context.merge(macro_option || { option_name => value }) }
 
       before do
         model.send macro, :company, **options
@@ -27,6 +31,21 @@ shared_examples_for 'an Active Record matcher' do
 
       it 'does not match if the options do not match' do
         should_not send(matcher_name, :company, option_name => negative)
+      end
+
+      instance_exec &block if block_given?
+    end
+  end
+
+  def self.raise_on_invalid_value(value, error_message)
+    value_description = value.is_a?(Proc) ? 'proc' : value.inspect
+
+    context "with #{value_description} value" do
+      let(:value) { value }
+
+      it 'raises an error' do
+        expect { should send(matcher_name, :company, option_name => value) }
+          .to raise_error ArgumentError, error_message
       end
     end
   end